Brings upstream's new optimization pipeline (`abf4911`), which is the reason for
the bump, plus 46 other commits: TP allocation and signal-handling fixes, CPU
cache pinning work, DSA-on-MLA fixed-size allocations, and a BlockSparseMLP
fused GPU->CPU handoff kernel.

Clean auto-merge. Only two files overlap our 19 commits -- eval/qbench.py and
eval/qbench/plot.py, where upstream added a KLD log-plot option and changed
colors -- and neither conflicted. Nothing upstream touched
exl3_gemm.cu/.cuh, modules/embedding.py or setup.py, so the two core changes we
carry (the EXL3_SM90_BARRIER build gate and block-quantized embedding loading)
are untouched.

Verified post-merge rather than assumed: every fake_quantize granularity still
runs (per_row, per_tensor, block, blockq, kshape, sym, sym_row), the symmetric
arms are still bit-identical to compressed_tensors' own encoder, and the diff
against v1.4.3 is the expected 11 files.

18 native sources changed upstream, so the extension needs rebuilding before
this is usable.

+++ exllamav3/modules/linear.py
@@ -811,7 +811,9 @@ def convert_exl3_group(
     for linear in linears:
         assert isinstance(linear.inner, LinearFP16), \
             "Inner layer is already quant type"
-        weights.append(linear.inner.get_weight_tensor().float())
+        # Checkpoint precision, usually still swapped to CPU: quantize_exl3_batch stages the
+        # upload through pinned memory and casts to fp32 on the device
+        weights.append(linear.inner.get_weight_tensor())
         biases.append(linear.inner.get_bias_tensor())
         linear.inner = None
